Fix disqus identifier (+ all the things)
- markdown was renaming .md to .html which is not nice (see the shittty
plugin I had to make to allow raw html files not wrapper by react
templates).
- So I dropped metalsmith-markdown (which hardcode the renaming) and
metalsmith-metallic for a homemade plugin to get back previous markdown
integration (hello anchor links).
- I also stripped all html tags from title to avoid anchor like
`-em-some-text-em-`.
- I also trimmed dash from escaped slug, that’s cool
- While I was on that, I improved the highlight (which close #319).
- Since I removed the hardcoding of .md to .html, I wrote a plugin to
rename files. So files are now renamed via a metalsmith plugin when we
ask for it, not by “you don’t know which plugin”.
- Since we are renaming after md to react/html I had to create a plugin
to add url metdata. Since it’s related to renaming, I found a stupid
and simple way to share the pattern md to html which is cool. Having
url field will allow us to have cleaner code so we can fix #377
- the url field can also be used by the rss plugin so let’s use it
- I also removed useless code from react-template plugin since renaming
is out (this was some legacy code from the original shitty plugin)
- while removing some code, I droppped debug module, because using log
to debug is so 90’
- No I can fix disqus identifier (close #377) with a smile

![](http://i.imgur.com/t0XHtgJ.gif)