Better assets caching
- css handled by webpack use extract plugin (to get an actual css file,
not the css embed in js file - make more sense for us while we are
style serving as some static content)
- assets version is not the latest commit hash. The reason is if a
commit affect a README file (for example), it make no sense at all to
change version of the assets, so instead of that we use a “content
hash” (that why I remove uglify in this commit, I need to find a way -
didn’t try yet) to get success build build from the same source that
give the same source - so content hash do not change because mangled
variables get renamed again (let’s say “a” to “b”…). During dev, no
hash used, for prod, webpack is run before metalsmith build so we can
easily share hashed name for the assets.
- React layouts have been splited from modules, to make maintenance
easier when you have to update all layouts (even if they are just react
class). Maybe we can find a better way for the import of the module,
this can still be improved, for now, we just import from
“../../modules/*”
- I removed some useless files (src/modules/base*) that are unused
since we use metalsmith-react